refactor (ontology): make interface an abstract class
