+++ src/cache/ontology-cache/resource-class-and-property-definitions.ts
@@ -1,14 +1,13 @@
-import { IClassAndPropertyDefinitions } from "../../models/v2/ontologies/IClassAndPropertyDefinitions";
+import { ClassAndPropertyDefinitions } from "../../models/v2/ontologies/ClassAndPropertyDefinitions";
 import { PropertyDefinition } from "../../models/v2/ontologies/property-definition";
 import { TypeGuard } from "../../models/v2/resources/type-guard";
 import { ResourceClassDefinitionWithPropertyDefinition } from "./resource-class-definition-with-property-definition";
-import { OntologyConversionUtil } from "../../models/v2/ontologies/OntologyConversionUtil";
 
 /**
  * Represents resource class definitions
  * and property definitions the resource classes have cardinalities for.
  */
-export class ResourceClassAndPropertyDefinitions implements IClassAndPropertyDefinitions {
+export class ResourceClassAndPropertyDefinitions extends ClassAndPropertyDefinitions {
 
     /**
      * Resource class definitions and their cardinalities.
@@ -21,7 +20,7 @@ export class ResourceClassAndPropertyDefinitions implements IClassAndPropertyDef
     properties: { [index: string]: PropertyDefinition };
 
     constructor(resClassDefs: { [index: string]: ResourceClassDefinitionWithPropertyDefinition }, propDefs: { [index: string]: PropertyDefinition }) {
-
+        super();
         this.classes = resClassDefs;
         this.properties = propDefs;
     }
@@ -30,7 +29,7 @@ export class ResourceClassAndPropertyDefinitions implements IClassAndPropertyDef
      * Gets all property definitions from the resource's entity info.
      */
     getAllPropertyDefinitions(): PropertyDefinition[] {
-        return OntologyConversionUtil.getAllEntityDefinitionsAsArray(this.properties);
+        return this.getAllEntityDefinitionsAsArray(this.properties);
     }
 
     /**
@@ -39,6 +38,6 @@ export class ResourceClassAndPropertyDefinitions implements IClassAndPropertyDef
      * @param type restriction to a certain property definition type.
      */
     getPropertyDefinitionsByType<T extends PropertyDefinition>(type: TypeGuard.Constructor<T>): T[] {
-        return OntologyConversionUtil.getEntityDefinitionsByTypeAsArray(this.getAllPropertyDefinitions(), type);
+        return this.getEntityDefinitionsByTypeAsArray(this.getAllPropertyDefinitions(), type);
     }
 }
