+++ src/models/v2/ontologies/read/read-ontology.ts
@@ -3,12 +3,11 @@ import { Constants } from "../../Constants";
 import { DateTimeStampConverter } from "../../custom-converters/date-time-stamp-converter";
 import { TypeGuard } from "../../resources/type-guard";
 import { ClassDefinition } from "../class-definition";
-import { IClassAndPropertyDefinitions } from "../IClassAndPropertyDefinitions";
-import { OntologyConversionUtil } from "../OntologyConversionUtil";
+import { ClassAndPropertyDefinitions } from "../ClassAndPropertyDefinitions";
 import { PropertyDefinition } from "../property-definition";
 
 @JsonObject("ReadOntology")
-export class ReadOntology implements IClassAndPropertyDefinitions {
+export class ReadOntology extends ClassAndPropertyDefinitions {
 
     @JsonProperty("@id", String)
     id: string = "";
@@ -28,7 +27,7 @@ export class ReadOntology implements IClassAndPropertyDefinitions {
      * Gets all class definitions from the ontology's entity info.
      */
     getAllClassDefinitions(): ClassDefinition[] {
-        return OntologyConversionUtil.getAllEntityDefinitionsAsArray(this.classes);
+        return this.getAllEntityDefinitionsAsArray(this.classes);
     }
 
     /**
@@ -37,14 +36,14 @@ export class ReadOntology implements IClassAndPropertyDefinitions {
      * @param type restriction to a certain class definition type.
      */
     getClassDefinitionsByType<T extends ClassDefinition>(type: TypeGuard.Constructor<T>): T[] {
-        return OntologyConversionUtil.getEntityDefinitionsByTypeAsArray(this.getAllClassDefinitions(), type);
+        return this.getEntityDefinitionsByTypeAsArray(this.getAllClassDefinitions(), type);
     }
 
     /**
      * Gets all property definitions from the ontology's entity info.
      */
     getAllPropertyDefinitions(): PropertyDefinition[] {
-        return OntologyConversionUtil.getAllEntityDefinitionsAsArray(this.properties);
+        return this.getAllEntityDefinitionsAsArray(this.properties);
     }
 
     /**
@@ -53,7 +52,7 @@ export class ReadOntology implements IClassAndPropertyDefinitions {
      * @param type restriction to a certain property definition type.
      */
     getPropertyDefinitionsByType<T extends PropertyDefinition>(type: TypeGuard.Constructor<T>): T[] {
-        return OntologyConversionUtil.getEntityDefinitionsByTypeAsArray(this.getAllPropertyDefinitions(), type);
+        return this.getEntityDefinitionsByTypeAsArray(this.getAllPropertyDefinitions(), type);
     }
 
 }
